+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2022.3] - 2022-10-X
+## [2022.3] - 2022-11-X
 
 ### Added
 
@@ -8,6 +8,8 @@
 - Documentation for drawing parameters
 - Support for Python 3.11
 - Function to convert a state into a 1D-numpy array
+- Progress bar for video creation
+- Callback for modifying axes during plotting a video
 
 ### Changed
 
@@ -16,6 +18,11 @@
 ### Removed
 
 - ParamServer and dictionary specification of drawing parameters
+
+### Fixed
+
+- Failing visualization of a trajectory at it's initial time step
+- Broken lanelet visualization with matplotlib >3.5
 
 ## [2022.2] - 2022-09-03